+++ javascipts/components/unhealthy-dinos.js
@@ -1,0 +1,6 @@
+// import { dinos } from './dino-data.js'
+
+// const unhealthyDinos = dinos.filter(dino => dino.health <= 49);
+// console.log(unhealthyDinos, "unhealth")
+
+// export { unhealthyDinos }
