+++ javascipts/main.js
@@ -1,0 +1,10 @@
+
+import { printDinos } from '../dino-data.js'
+// import { unhealthyDinos } from './unhealthy-dinos.js'
+
+const init = () => {
+    dinos()
+    printDinos();
+}
+
+init()
