--- main.js
@@ -1,8 +1,0 @@
-
-import { printDinos } from './dino-data.js'
-
-const init = () => {
-    printDinos();
-}
-
-init()
